+++ Source/PropertyType/Handle.js
@@ -6,8 +6,8 @@ Handle.REG_EX = /^([\-0-9\.]+)\,([\-0-9\.]+)$/;
 Handle.fromString = function(literal) {
     var handle = new Handle();
     if (literal.match(Handle.REG_EX)) {
-        handle.x = parseInt(RegExp.$1);
-        handle.y = parseInt(RegExp.$2);
+        handle.x = parseFloat(RegExp.$1);
+        handle.y = parseFloat(RegExp.$2);
     }
 
     return handle;
